Create job card details component
