+++ src/components/Jobs/Jobs.js
@@ -1,18 +1,10 @@
 import React from 'react';
 import styled from 'styled-components';
-import { NavLink, Link } from 'react-router-dom';
-import { Header, Button, Footer } from '../~common';
-import {
-	small_space,
-	dark,
-	white,
-	boxShadow,
-	medium_space,
-	borderRadius,
-	primary_color,
-	ContainerStyled
-} from '../~styles/reuseables';
+import { NavLink } from 'react-router-dom';
+import { Header, Footer } from '../~common';
+import { white, medium_space, ContainerStyled } from '../~styles/reuseables';
 import { Job } from './Job';
+import { JobDetails } from './JobDetails';
 
 const jobs = [
 	{
@@ -91,24 +83,14 @@ export const Jobs = () => {
 						<JobListContainer>{jobs.map(job => <Job {...job} />)}</JobListContainer>
 
 						<RightAside>
-							<h3>Job details</h3>
-							<CompanyDetailsWrapper>
-								<div>
-									<img src="https://img.icons8.com/color/48/000000/google-logo.png" alt="google" />
-								</div>
-								<ul>
-									<li>Google inc</li>
-									<li>Location</li>
-									<li>Company size</li>
-									<li>Linked in</li>
-									<li>Website</li>
-								</ul>
-							</CompanyDetailsWrapper>
-
-							<ActionWrapper>
-								<Next buttonText="Next" type="button" onClick={() => console.log('ok')} />
-								<SuperLike buttonText="Like" type="button" onClick={() => console.log('ok')} />
-							</ActionWrapper>
+							<JobDetails
+								imgSrc="https://img.icons8.com/color/48/000000/google-logo.png"
+								companyName="Google"
+								location="Lagos Nigeria"
+								companySize="10000"
+								linkedIn="linkedin/google"
+								website="https://www.google.com/"
+							/>
 						</RightAside>
 					</MainWrapper>
 				</ContainerStyled>
@@ -144,57 +126,10 @@ const MainWrapper = styled.div`
 	justify-content: space-between;
 `;
 
-const RightAside = styled.aside`
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	justify-content: space-between;
-	background: ${white};
-	min-height: 600px;
-	width: 350px;
-	box-shadow: ${boxShadow};
-	padding-top: ${medium_space};
-	border-radius: ${borderRadius};
-	padding: ${small_space};
-
-	a {
-		color: ${dark};
-
-		&:hover {
-			color: ${primary_color};
-		}
-	}
-`;
+const RightAside = styled.aside``;
 
 const JobListContainer = styled.section`
 	width: 900px;
 	display: flex;
 	flex-wrap: wrap;
-`;
-
-const CompanyDetailsWrapper = styled.div`
-	text-align: center;
-
-	ul {
-		list-style: none;
-		padding: ${small_space};
-
-		li {
-			margin-bottom: ${small_space};
-		}
-	}
-`;
-
-const SuperLike = styled(Button)``;
-
-const Next = styled(Button)``;
-
-const ActionWrapper = styled.div`
-	display: flex;
-
-	button {
-		width: 150px;
-		margin: 10px;
-		font-weight: 400;
-	}
 `;
